+++ packages/flutter_adaptive_cards_fs/test/utils/test_utils.dart
@@ -240,17 +240,11 @@ class Blue8x8Image {
 Widget getTestWidgetFromPath({
   required String path,
   Key? key,
-  Function(String)? onOpenUrl,
-  Function(String)? onOpenUrlDialog,
+  void Function(OpenUrlActionInvoke invoke)? onOpenUrl,
+  void Function(OpenUrlDialogActionInvoke invoke)? onOpenUrlDialog,
   void Function(SubmitActionInvoke invoke)? onSubmit,
   void Function(ExecuteActionInvoke invoke)? onExecute,
-  Function(
-    String id,
-    dynamic value,
-    DataQuery? dataQuery,
-    RawAdaptiveCardState cardState,
-  )?
-  onChange,
+  void Function(InputChangeInvoke invoke)? onChange,
   Map? initData,
   HostConfigs? hostConfigs,
   bool listView = false,
@@ -280,17 +274,11 @@ Widget getTestWidgetFromMap({
   required Map<String, dynamic> map,
   required String title,
   Key? key,
-  Function(String)? onOpenUrl,
-  Function(String)? onOpenUrlDialog,
+  void Function(OpenUrlActionInvoke invoke)? onOpenUrl,
+  void Function(OpenUrlDialogActionInvoke invoke)? onOpenUrlDialog,
   void Function(SubmitActionInvoke invoke)? onSubmit,
   void Function(ExecuteActionInvoke invoke)? onExecute,
-  Function(
-    String id,
-    dynamic value,
-    DataQuery? dataQuery,
-    RawAdaptiveCardState cardState,
-  )?
-  onChange,
+  void Function(InputChangeInvoke invoke)? onChange,
   Map? initData,
   HostConfigs? hostConfigs,
   bool listView = false,
@@ -332,26 +320,17 @@ Widget getTestWidgetFromMap({
             InheritedAdaptiveCardHandlers(
               // this a test so we can look at this later
               // ignore: inference_failure_on_collection_literal
-              onOpenUrl: onOpenUrl ?? (String _) => {},
+              onOpenUrl: onOpenUrl ?? (_) {},
               // this a test so we can look at this later
               // ignore: inference_failure_on_collection_literal
-              onOpenUrlDialog: onOpenUrlDialog ?? (String _) => {},
+              onOpenUrlDialog: onOpenUrlDialog ?? (_) {},
               // this a test so we can look at this later
               // ignore: inference_failure_on_collection_literal
               onSubmit: onSubmit ?? (_) {},
               // this a test so we can look at this later
               // ignore: inference_failure_on_collection_literal
               onExecute: onExecute ?? (_) {},
-              onChange:
-                  onChange ??
-                  (
-                    String id,
-                    dynamic value,
-                    DataQuery? dataQuery,
-                    RawAdaptiveCardState cardState,
-                    // this a test so we can look at this later
-                    // ignore: inference_failure_on_collection_literal
-                  ) => {},
+              onChange: onChange ?? (_) {},
               child: adaptiveCard,
             ),
           ),